Fix no javascript subscription POST, remove test I18n files.
Added hidden inputs for nonce and action. They are present in advanced.php but
for some reason weren't added to basic.php. Due to this the basic form wouldn't
run the function.
Made the function redirect to previous url if present, if not redirect to homepage.
Also removed some leftover testing I18n that I accidentally commited.